Enable bootarg in order to force use of RGB colorspace
Some old TVs (tested with Philips) report incorrectly HDMI requests to change colorspace.
The result is an image flooded with green/magenta colors.

This workaround enables a bootarg hdmitx=forcergb in order to ignore the HDMI request
from the TV and use always RGB colorspace. I've only created the bootarg because
the source code to force the colorspace was already there but surprisingly was not being used.